First transfer of working tile code into the strato
TODO:
watch for task stack overflow
psram thread safe?
speed of execution

+++ BB3/App/etc/geo_calc.h
@@ -15,7 +15,7 @@
 #define MAP_W   200
 #define MAP_H   200
 
-void geo_get_steps(int32_t lat, uint8_t zoom, int32_t * step_x, int32_t * step_y);
+void geo_get_steps(int32_t lat, uint16_t zoom, int32_t * step_x, int32_t * step_y);
 void geo_to_pix(int32_t lon, int32_t lat, uint8_t zoom, int32_t g_lon, int32_t g_lat, int16_t * x, int16_t * y);
 void geo_to_pix_w_h(int32_t lon, int32_t lat, uint8_t zoom, int32_t g_lon, int32_t g_lat, int16_t * x, int16_t * y, int16_t w, int16_t h);
 void geo_get_topo_steps(int32_t lat, int32_t step_x, int32_t step_y, int16_t * step_x_m, int16_t * step_y_m);
